ignore carlbot

+++ bot.js
@@ -23,6 +23,7 @@ process.on('unhandledRejection', (error) => {
 
 bot.on("message", (message) => {
   if (message.channel.id === "804351457054031902") {
+    if (message.author.id === "235148962103951360") return;
     if (!message.content.startsWith("!suggest")) {
       message.delete();
     }
